Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/html/browsers/windows/embedded-opener-expected.txt


FAIL opener and embedded documents; using window.open() assert_equals: opener before setting it to null via directly invoking the getter expected object "[object Window]" but got null
